Enhances test workflow for dev Symfony versions
Updates the test workflow to correctly handle "-dev" Symfony versions.

Sets the composer minimum-stability to "dev" and prefers stable dependencies
when testing against development versions of Symfony. This ensures that the
latest dependencies are used during testing of development versions and also
runs update with --prefer-lowest only when not on a dev branch.